+++ src/styles/defaultColor.js
